Onboarding: Generalize `PercentageField` component
